+++ forum-service/src/main/java/telran/java51/post/controller/PostController.java
@@ -1,5 +1,6 @@
 package telran.java51.post.controller;
 
+import java.security.Principal;
 import java.util.List;
 
 import org.springframework.http.HttpStatus;
@@ -14,6 +15,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
+import telran.java51.accounting.dto.UserDto;
+import telran.java51.accounting.dto.exceptions.UserForbiddenException;
+import telran.java51.accounting.service.UserService;
 import telran.java51.post.dto.DatePeriodDto;
 import telran.java51.post.dto.NewCommentDto;
 import telran.java51.post.dto.NewPostDto;
@@ -25,11 +29,14 @@ import telran.java51.post.service.PostService;
 @RequestMapping("/forum")
 public class PostController {
 
+	final UserService userService;
 	final PostService postService;
+	List <String> roles=List.of("ADMINISTRATOR", "MODERATOR");
+	
 
 	@PostMapping("/post/{author}")
-	public PostDto addNewPost(@PathVariable String author, @RequestBody NewPostDto newPostDto) {
-		return postService.addNewPost(author, newPostDto);
+	public PostDto addNewPost(Principal principal, @PathVariable String author, @RequestBody NewPostDto newPostDto) {
+		return postService.addNewPost(principal.getName(), newPostDto);
 	}
 
 	@GetMapping("/post/{id}")
@@ -38,19 +45,33 @@ public class PostController {
 	}
 
 	@DeleteMapping("/post/{id}")
-	public PostDto removePost(@PathVariable String id) {
-		return postService.removePost(id);
+	public PostDto removePost(Principal principal, @PathVariable String id) {
+		
+		if(checkRole(principal, roles.subList(0,2)) || checkLogin(principal.getName(), postService.findPostById(id).getAuthor())) {
+			return postService.removePost(id);
+		} else {
+			
+			throw new UserForbiddenException();
+		}
+	
 	}
 
 	@PutMapping("/post/{id}")
-	public PostDto updatePost(@PathVariable String id, @RequestBody NewPostDto newPostDto) {
-		return postService.updatePost(id, newPostDto);
+	public PostDto updatePost(Principal principal, @PathVariable String id, @RequestBody NewPostDto newPostDto) {
+		
+		if(checkRole(principal, roles.subList(0,1)) || checkLogin(principal.getName(), postService.findPostById(id).getAuthor())) {
+			return postService.updatePost(id, newPostDto);
+		} else {
+			
+			throw new UserForbiddenException();
+		}
+		
 	}
 
 	@PutMapping("/post/{id}/comment/{author}")
-	public PostDto addComment(@PathVariable String id, @PathVariable String author,
+	public PostDto addComment(Principal principal, @PathVariable String id, @PathVariable String author,
 			@RequestBody NewCommentDto newCommentDto) {
-		return postService.addComment(id, author, newCommentDto);
+		return postService.addComment(id, principal.getName(), newCommentDto);
 	}
 
 	@PutMapping("/post/{id}/like")
@@ -75,4 +96,20 @@ public class PostController {
 		return postService.findPostsByPeriod(datePeriodDto);
 	}
 
+	
+	
+	private boolean checkRole(Principal principal, List <String> roles) {
+		
+		UserDto userDto = userService.findById(principal.getName());
+		
+		return roles.stream().anyMatch(x->userDto.getRoles().contains(x));
+
+	}
+	
+	private boolean checkLogin(String name, String login) {
+		
+		return name.equalsIgnoreCase(login);
+	}
+	
+	
 }
